計算ロジックを静的関数に移動
