+++ app/src/main/java/com/oukoda/svinfocompose/view/component/speedrank/SpeedRankListItem.kt
@@ -1,0 +1,2 @@
+package com.oukoda.svinfocompose.view.component.speedrank
+
